requirements.txt
google-adk
google-cloud-aiplatform 